Add Rails 4 support
Quite a few things needed to be addressed to make this happen:

Bump AR version in gemspec.

The test at the following line was failing due to the `except` option.
"email" was replaced with "name".
https://github.com/vigetlabs/serialize_with_options/blob/d792a472431c72b93dd00b14e7654a8f2c7ebb89/test/serialize_with_options_test.rb#L184

The `encode` method within ActiveSupport calls `dup` on `options`.
When using `to_json`, we were allowing the Symbol for `set` to be
passed to this method. `dup` cannot be called on Symbol, so `to_json`
had to be added back as a method that we parse the options for. Here's
the offending line in Rails:
https://github.com/rails/rails/blob/4-1-6/activesupport/lib/active_support/json/encoding.rb#L34